requirements.txt
torchvision
efficientnet-pytorch
albumentations
opencv-python
numpy
matplotlib
scipy
pillow
typing
customtkinter
flask
pytube